next attempt on macos

+++ gui/src/scanHelper.cc
@@ -40,9 +40,9 @@ bool Scan::setupScan(std::string globalConfFile, std::string scanConfFile) {
             if(!m_scan_is_time_based && Counter == "wrongPara")
                 return scanError("To run a scan based on a number of events, \"eventCounter\" needs to be specified in section"+std::to_string(i));
             if(name == "wrongPara" || param == "wrongPara"
-               || (std::numeric_limits<double>::epsilon()>std::abs(start - std::numeric_limits<double>::min()))
-               || (std::numeric_limits<double>::epsilon()>std::abs(stop - std::numeric_limits<double>::min()))
-               || (std::numeric_limits<double>::epsilon()>std::abs(step - std::numeric_limits<double>::min())))
+               || (std::numeric_limits<double>::epsilon()>std::abs(double(start - std::numeric_limits<double>::min())))
+               || (std::numeric_limits<double>::epsilon()>std::abs(double(stop - std::numeric_limits<double>::min())))
+               || (std::numeric_limits<double>::epsilon()>std::abs(double(step - std::numeric_limits<double>::min()))))
                 return scanError("Scan section "+std::to_string(i)+" is incomplete -> Please check");
 
             defaultConf->SetSection("");
